demo/requirements.txt
# Core Python dependencies
python-dotenv>=1.0.0

# OpenAI for LLM functionality
openai>=1.0.0

# HTTP client for API calls
aiohttp>=3.8.0

# Google Cloud dependencies for PubSub
google-cloud-pubsub>=2.18.0
google-auth>=2.17.0

# Process monitoring
psutil>=5.9.0

# Textual for TUI (main.py)
textual>=0.40.0
textual-dev>=1.2.0

# LiveKit for voice calls (call.py)
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]==1.0.17
livekit-plugins-noise-cancellation~=0.2

# Pydantic for data validation
pydantic>=2.0.0
pydantic-core>=2.0.0

# Audio processing (for call.py)
pyaudio>=0.2.11

# Twilio for SMS/WhatsApp (sample.py)
twilio>=8.0.0
